+++ assets/javascripts/kitten/components/form/field/field.stories.js
@@ -3,6 +3,7 @@ import { storiesOf } from '@storybook/react'
 import { withKnobs, text, number, boolean } from '@storybook/addon-knobs'
 import {
   FieldInputExample,
+  FieldPasswordExample,
   FieldRadioButtonSetExample,
   FieldSelectExample,
 } from './field.examples'
@@ -36,6 +37,22 @@ storiesOf('Form/Field', module)
           error={boolean('Error?', false)}
           errorMessage={text('Error', 'Error message…')}
           limit={number('Limit', undefined)}
+          unit={text('Unit', undefined)}
+        />
+      </StoryGrid>
+    )
+  })
+  .add('with password', () => {
+    return (
+      <StoryGrid>
+        <FieldPasswordExample
+          id={text('ID', 'input')}
+          label={text('Label', 'Label')}
+          tooltip={text('Tooltip', null)}
+          tooltipId={text('Tooltip ID', 'tooltip')}
+          placeholder={text('Placeholder', 'Placeholder…')}
+          error={boolean('Error?', false)}
+          errorMessage={text('Error', 'Error message…')}
         />
       </StoryGrid>
     )